program that multiplies two numbers

+++ 0x0C-more_malloc_free/101-mul.c
@@ -19,7 +19,7 @@ int isDigit(char s)
 
 int _atoi(char *s)
 {
-	unsigned int num = 0;
+	unsigned long int num = 0;
 	int sign = 1;
 
 	do {
